arm: tegra: Convert SOC_THERM to platform driver
Converted the SOC_THERM library to a platform driver in order for it
to be more in line with the upstream Linux kernel.

In the (rare) case of different order of initialization of soctherm,
it is possible that the throttle_depth string could be NULL. Added a
NULL check before doing strcmp on throttle_depth to prevent crash.

Bug 200004664

Change-Id: Ib0bda73651b73a69f704f77d6474344f58c8c69d
Signed-off-by: Diwakar Tundlam <dtundlam@nvidia.com>
Reviewed-on: http://git-master/r/437436
Reviewed-by: Automatic_Commit_Validation_User