arm64: tegra132: enable PSCI v0.2 support
Enable PSCI as the CPU "enable-method" on T132. This means we also have
to support secondary CPU boot, cpu idle and LP0 using PSCI calls. The
flow controller programming has to be moved to EL3 monitor and the kernel
code can be removed. The extra kernel code which was added to support a
spin-table approach has been removed in order to avoid any confusion. Now
that we have switched to PSCI, there should be no need to go back to
spin-tables again.

The monitor code already supports v0.2, so we can safely enable support
in the kernel as well.

Change-Id: Ia3b4abec49c3eb49cdffae3a598eeacec7fbe74d
Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>
Reviewed-on: http://git-master/r/606086
Reviewed-by: Peng Du <pdu@nvidia.com>
GVS: Gerrit_Virtual_Submit