regulator: core: add support to profile timing on voltage change
Add support for the time profiling on voltage change. This keeps track
of the time taken on voltage change on different time ranges and also
minimum/maximum time needed for voltage change.

It supports the two sysfs as
	set_voltage_time_profile_enable
	timing_profile

The profiling can be enable through sysfs when it is needed as
	echo 1 > /sys/kernel/debug/regulator/vdd-core/timing_profile

Then time time for voltage change is find from following sysfs
	cat /sys/kernel/debug/regulator/vdd-core/set_voltage_time_profile

This outputs the message as:
----
min_time: 0
max_time: 8000
Time(us):  Count
0 - 25 : 43
1000 - 1100 : 65
2000 - 2200 : 30
3000 - 3200 : 3
4000 - 4200 : 2
5000 - 5500 : 1
7000 - 7500 : 2
8000 - 8500 : 1
----
Here min_time is minimum time taken for voltage change,
max_time is maximum time taken for voltage change and
other data are the number of times, the voltage change
occurred on that time range i.e. voltage change completed
65 times on the range of 1000 to 1100us.

With above, it can observe that most of time, voltage change is
happening on 1000us and it took more than 8ms once.

bug 1409234

Change-Id: Iaabb6d81c23b265c0601a2fe9a9f7c143113d866
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/670226
GVS: Gerrit_Virtual_Submit